requirements.txt
flask
requests
sqlalchemy
pymysql
python-dotenv
mysql-connector-python
